Create Vision Statement
Product Vision Statement:

Here at HairbyBri, we're for millennial college students who are looking for creative and unique hairstyles for any event or just everyday basis. HairByBri is a unique hair care booking site that focus on making sure our customers are satisfied with the hairstyles and services they will be receiving unlike other hairstylists who only have social media pages that you can direct message them through to book an service needed. With the HairByBri booking site you will be able to see all the services offered with pictures, the cost of the service and the time it takes to complete your selected service.